Migrate DamageModifier (ArmorModifier) from float to integer percentage.

+++ OpenRA.Game/Traits/TraitsInterfaces.cs
@@ -149,7 +149,7 @@ namespace OpenRA.Traits
 	}
 
 	public interface IRenderModifier { IEnumerable<IRenderable> ModifyRender(Actor self, WorldRenderer wr, IEnumerable<IRenderable> r); }
-	public interface IDamageModifier { float GetDamageModifier(Actor attacker, WarheadInfo warhead); }
+	public interface IDamageModifier { int GetDamageModifier(Actor attacker, WarheadInfo warhead); }
 	public interface ISpeedModifier { int GetSpeedModifier(); }
 	public interface IFirepowerModifier { float GetFirepowerModifier(); }
 	public interface ILoadsPalettes { void LoadPalettes(WorldRenderer wr); }
